+++ examples/php-code-quality-package/src/Rule.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Acme\QualityKit;
+
+class Rule
+{
+}